backend/mcp-aegis/requirements.txt
mcp>=0.1.0
flask>=2.0.0
aiohttp>=3.8.0
openai>=0.27.0
hypercorn>=0.14.0
asgiref>=3.5.0
nest_asyncio>=1.5.5
python-dotenv>=0.19.0